requirements.txt
duckdb
polars
pyarrow
python-telegram-bot
psutil
